+++ adapters/built/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0
+
+* Update mapper adapter contract to latest
+
 ## 0.0.1
 
 * Initial support for BuiltList, BuiltSet, BuiltMap types